# product-api/swagger.yaml
basePath: /
consumes:
    - application/json
definitions:
    GenericError:
        description: GenericError GenericError GenericError generic error
        properties:
            Message:
                description: generic message for most of the errors
                type: string
        type: object
        x-go-package: github.com/zakisk/microservice/product-api/sdk/models
    Product:
        description: Product Product Product model class for product
        properties:
            description:
                description: description of product
                type: string
                x-go-name: Description
            id:
                description: id of product
                format: int64
                type: integer
                x-go-name: ID
            name:
                description: name of product
                minLength: 2
                type: string
                x-go-name: Name
            price:
                description: price of product
                format: float
                minimum: 1
                type: number
                x-go-name: Price
            sku:
                description: unique identifier of the product
                pattern: '[a-z]+-[a-z]+-[a-z]+'
                type: string
                x-go-name: SKU
        required:
            - id
            - name
            - price
            - sku
        type: object
        x-go-package: github.com/zakisk/microservice/product-api/sdk/models
    ValidationError:
        description: ValidationError ValidationError ValidationError ValidationError is a slice of json validation errors returned by validator
        properties:
            Messages:
                description: messages
                items:
                    type: string
                type: array
        type: object
        x-go-package: github.com/zakisk/microservice/product-api/sdk/models
host: localhost:9090
info:
    description: Documentation of our awesome API.
    title: awesome.
    version: 1.0.0
paths:
    /products:
        get:
            description: Returns a list of products
            operationId: listProducts
            parameters:
                - description: |-
                    Currency used when returning the price of the product,
                    when not specified currency is returned in GBP.
                  in: query
                  name: Currency
                  type: string
            responses:
                "200":
                    $ref: '#/responses/productResponse'
                "500":
                    $ref: '#/responses/internalServerError'
            tags:
                - products
        put:
            operationId: updateProduct
            parameters:
                - description: |-
                    Product data structure to Update or Create.
                    Note: the id field is ignored by update and create operations
                  in: body
                  name: Body
                  required: true
                  schema:
                    $ref: '#/definitions/Product'
            responses:
                "201":
                    $ref: '#/responses/noContentResponse'
            tags:
                - products
    /products/{id}:
        delete:
            description: Deletes a product from database of given id
            operationId: deleteProduct
            parameters:
                - description: The id of the product for which the operation relates
                  format: int64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
            responses:
                "201":
                    $ref: '#/responses/noContentResponse'
                "404":
                    $ref: '#/responses/notFound'
                "500":
                    $ref: '#/responses/internalServerError'
            tags:
                - products
        get:
            description: Returns a list of products
            operationId: listSingleProducts
            responses:
                "200":
                    $ref: '#/responses/productResponse'
                "404":
                    $ref: '#/responses/notFound'
                "500":
                    $ref: '#/responses/internalServerError'
            tags:
                - products
produces:
    - application/json
responses:
    errorResponse:
        description: Generic error message returned as a string
        schema:
            $ref: '#/definitions/GenericError'
    errorValidation:
        description: Validation errors defined as an array of strings
        schema:
            $ref: '#/definitions/ValidationError'
    internalServerError:
        description: Internal Server error when something went wrong at server end
        schema:
            $ref: '#/definitions/GenericError'
    noContentResponse:
        description: No content is returned by this API endpoint
    notFound:
        description: ""
        schema:
            $ref: '#/definitions/GenericError'
    productResponse:
        description: Data structure representing a single product
        schema:
            $ref: '#/definitions/Product'
    productsResponse:
        description: A list of products
        schema:
            items:
                $ref: '#/definitions/Product'
            type: array
schemes:
    - http
securityDefinitions:
    basic:
        type: basic
swagger: "2.0"